Add some additional parsing tests
Trying to pin down exactly what the problem is for #62. This adds
some parsing tests to  help me understand how Edsger is parsing
the expressions that lead to the bug in #62.

These pass, which makes me think that the problem may not be
parsing _per se_, but instead some inconsistency in how we're
internally representing expressions.