Extend testing for bug #62
@nathan-beneke had added a test that failed, demonstrating the issue with #62.
This makes the test pass, _but not by fixing the bug_. Our expressions seem
to have a lot of internal quoting, which seems wrong to me, e.g.,

```
'(:not '(:and '(:and p p) q))
```

instead of just

```
'(:not (:and (:and p p) q))
```

If I remove the internal quoting, the test passes. I think this is because
the internal quoting creates a bunch of `quote` symbols that then don't
link up correctly in the recursive matching.

If this is right, then the problem may be that some interaction between the UI
and the parsing may be creating these extra quotes when we run the system,
causing the bug to appear.