Fix bugs #62 and #91
The problem appears to be that `check-match-recursive` had an unnecessary check that expressions were non-literals which used `list?`, which returned `false` if handed a vector. The parser we're using returns vectors, so complex expressions are parsed to nested vectors. These fail the `list?` test, and we fail to recurse appropriately. Simply removing the test appears to both fix the problem and still pass all the tests.

I also extended several of the tests to cover more cases.

Closes #62
Closes #91